requirements.txt
XlsxWriter==1.3.7
Flask_Bcrypt==0.7.1
Flask_WTF==0.14.3
Jinja2==2.10
pdfkit==0.6.1
Flask==1.1.2
itsdangerous==1.1.0
Flask_SQLAlchemy==2.5.1
WTForms==2.3.3
Flask_Mail==0.9.1
Flask_Login==0.5.0
Pillow==8.2.0
email-validator==1.1.2
pandas==1.1.4